Make the Element.getAttribute*() methods generic.
The Element.getAttribute collection of methods now automatically cast
their arguments on the expected type of the receiver variable:

String s = node.getAttribute("foo");

Expects that the attribute is a string. However if the attribute is not
a string, a class cast exception still occurs. This only moves the cast
inside the de method getAttribute(), no more.

In contrast the:

String s = node.getAttribute("foo",String.class);

Call will effectivelly test if the attribute is of the correct kind.
And this method is cleaner since it does not require a cast.