Corrected a bug in DefaultGraph.
The bug prevented the "autoCreate" feature to work. Since the
introduction of events time and source, the old addEdge_() was kept.
However it tried to create the event time, say 1, for the edge then if the
nodes of the edge were not yet created it created the nodes at time 2
and 3, and only
then created the edge with the old time 1. This yield situations where
the stream of events would send events in the incorect order ignoring
the event one, since this is the mechanism used to do synchronisation
between graphs.

A resurgence of this bug appeared with the viewer, the edge being
created in the graph, but not in the viewer since the synchronisation
between the viewer graph and the original graph was impossible.