Added the (long awaited) auto-stop to the layout.
The layout when started as an automatic separate process will
no stop if the global energy of the force-directed algorithm
does not vary. As for some layouts the global energy may always
vary a little it stops when the stabilization (difference between
the current energy and the average energy from the previous 256
iterations) is above a given threshold. The stabilization is a
number between 0 and 1, with 1 meaning fully stabilized. The
default limit is 0.9. One can change it by putting the
"layout.stabilization-limit" attribute on the graph with a
number value between 0 and 1. If it is set to 0, the auto-stop
feature will be disabled and the layout process will run forever.
Naturally, as soon as some change occurs in the graph, the layout
process wakes up.