Corrected a bug preventing sprites to appear if Graph.display() was called after their creation.
A subtle bug: when display is called late and the graph is already
created, with some sprites, the graph is "replayed" with all its
attributes.

Sprites are always stored as attributes on the graph. There is an
attribute that created the sprite, and as many attributes as needed for
each sprite own attributes. They all are stored in the graph.

However, attributes are replayed in an arbitrary order, therefore
attributes specifying that sprites must appear, and attributes
specifying the sprites own attributes may come in any order, if a change
of an attribute on sprites was occuring before the sprite creation was
encountered, it was ignored. This is no more the case.