The CSS parser did not issued errors for some invalid ui.style values.
When using ui.style, the parser was started using the "style" rule, not
the "start" rule that parses a whole stylesheet. The "style" rule
stops when the thing it reads is not a style property, and merely
ignore (does not recognize) the remaining parts of the things to
parse.

Added a "styleStart" rule that expect <EOF> at the end, so that
if something unknown is encountered, the parser emit an error.