Correction of another bug in the viewer tied to swing matrices use.
When putting the viewer in a GUI, the position of the component was not
used to compute the transformation matrices (the Graphics2D transforms
are used by Swing at certain times to move a whole component, the code
of the viewer used to reset the transformation to identity before
setting up the rendering space). The visibility check also had to be
modified to take this into account.