Fixes #57
The problem posed by this bug is double.

First it evidenced a bug in the stylesheet : styles do not accumulate
correctly when the style groups do not yet contain any element. Indeed
a style group is made to be updated, that is, you can specify several
times a style rule in the style sheet with different (or the same) style
definition. The style referenced by the rule will be augmented and
merged. However when the style does not reference any element, there
was bug when doing this. Since the CSS exists in GraphStream this never
happened !!!

This fix, corrects this bug, however there is another bug evidenced :
The code shown by Stefan sends twice the attribute "ui.stylesheet" to
the viewer. This bug will be corrected in a future commit.