Fixes #48.
The graph was not replayed in the Viewer when using a direct policy like
GRAPH_IN_SWING_THREAD.

The viewer works using events comming from the main graph. When the
graph is populated with nodes and edge and the viewer is attached on
it at a later time, the already present elements was not sent anew
to the viewer when using the GRAPH_IN_SWING_THREAD policy.