The SpringBox layout is now usable as a base class.
The old GraphStream SpringBox graph layout is now a cut into a
BarnesHutLayout class that serves as base for several force-based layout
algorithms and a new layout.implementations package contains a SpringBox
class that inherit BarnesHutLayout.

Nothing changed in the layout in itself, excepted it is now possible to
create new force-based layouts more easily using BarnesHutLayout as a
base.