AdjacencyListGraph does not check existence of a node/edge before shrinking array
When calling `AdjacencyListGraph.removeNode(n)` with a node `n` not present in the graph, it proceeds to call `removeNodeCallback(n)`. Despite `nodeMap.remove(...)` being a no-op in this case, the `nodeArray` loses the element in `nodeArray[node.getIndex()]`, which may actually also cause an `ArrayOutOfBoundsException`. The proposed change fixes this by checking if `nodeMap.remove(...)` actually found the node or not.

The same was done for edges.